Add NuGet provider in Frosting
This will make it possible to install NuGet tools in Frosting.
I also cleaned up the code a bit to make it possible to initialize
a NuGet module without providing a Cake configuration.

Closes #2933
