+++ src/Cake.NuGet.Tests/Fixtures/NuGetModuleFixture.cs
@@ -4,8 +4,6 @@
 
 using System;
 using Cake.Core.Composition;
-using Cake.Core.Configuration;
-using Cake.Testing;
 using NSubstitute;
 
 namespace Cake.NuGet.Tests.Fixtures
@@ -14,13 +12,11 @@ namespace Cake.NuGet.Tests.Fixtures
     {
         public ICakeContainerRegistrar Registrar { get; }
         public ICakeRegistrationBuilder Builder { get; }
-        public FakeConfiguration Configuration { get; }
 
         public NuGetModuleFixture()
         {
             Registrar = Substitute.For<ICakeContainerRegistrar>();
             Builder = Substitute.For<ICakeRegistrationBuilder>();
-            Configuration = new FakeConfiguration();
 
             Registrar.RegisterType<T>().Returns(Builder);
             Builder.As(Arg.Any<Type>()).Returns(Builder);
@@ -31,7 +27,7 @@ namespace Cake.NuGet.Tests.Fixtures
 
         public NuGetModule CreateModule()
         {
-            return new NuGetModule(Configuration);
+            return new NuGetModule();
         }
     }
 }
