+++ src/Cake.NuGet.Tests/Unit/NuGetLoadDirectiveProviderTests.cs
@@ -2,11 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using Cake.Core.Diagnostics;
 using Cake.Core.Packaging;
 using Cake.NuGet.Tests.Fixtures;
-using Cake.Testing.Xunit;
 using Xunit;
 
 namespace Cake.NuGet.Tests.Unit
