+++ src/Cake.NuGet/NuGetModule.cs
@@ -4,7 +4,6 @@
 
 using System;
 using Cake.Core.Composition;
-using Cake.Core.Configuration;
 using Cake.Core.Packaging;
 using Cake.Core.Scripting.Processors.Loading;
 
@@ -16,17 +15,6 @@ namespace Cake.NuGet
     /// </summary>
     public sealed class NuGetModule : ICakeModule
     {
-        private readonly ICakeConfiguration _config;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="NuGetModule"/> class.
-        /// </summary>
-        /// <param name="config">The config.</param>
-        public NuGetModule(ICakeConfiguration config)
-        {
-            _config = config ?? throw new ArgumentNullException(nameof(config));
-        }
-
         /// <inheritdoc/>
         public void Register(ICakeContainerRegistrar registrar)
         {
@@ -35,39 +23,12 @@ namespace Cake.NuGet
                 throw new ArgumentNullException(nameof(registrar));
             }
 
-            // NuGet content resolver.
-            registrar.RegisterType<NuGetContentResolver>()
-                .As<INuGetContentResolver>()
-                .Singleton();
+            registrar.RegisterType<NuGetPackageInstaller>().As<INuGetPackageInstaller>().As<IPackageInstaller>().Singleton();
+            registrar.RegisterType<NuGetContentResolver>().As<INuGetContentResolver>().Singleton();
+            registrar.RegisterType<NuGetLoadDirectiveProvider>().As<ILoadDirectiveProvider>().Singleton();
 
-#if !NETCORE
-            // Load directive provider.
-            registrar.RegisterType<NuGetLoadDirectiveProvider>()
-                .As<ILoadDirectiveProvider>()
-                .Singleton();
-#endif
-
-            // URI resource support.
-            if (!bool.TryParse(_config.GetValue(Constants.NuGet.UseInProcessClient) ?? bool.TrueString, out bool useInProcessClient) || useInProcessClient)
-            {
-#if NETCORE
-                // Load directive provider.
-                registrar.RegisterType<NuGetLoadDirectiveProvider>()
-                    .As<ILoadDirectiveProvider>()
-                    .Singleton();
-#endif
-                registrar.RegisterType<Install.NuGetPackageInstaller>()
-                    .As<INuGetPackageInstaller>()
-                    .As<IPackageInstaller>()
-                    .Singleton();
-            }
-            else
-            {
-                registrar.RegisterType<NuGetPackageInstaller>()
-                    .As<INuGetPackageInstaller>()
-                    .As<IPackageInstaller>()
-                    .Singleton();
-            }
+            registrar.RegisterType<InProcessInstaller>().Singleton();
+            registrar.RegisterType<OutOfProcessInstaller>().Singleton();
         }
     }
 }
